refactor(DateTimePreviewContent): Rename to DateTimePreview
